PhoneSpecs is now a container, added phonespecs component and types
