+++ src/components/PhoneSpecs/BasicInfo.tsx
@@ -2,9 +2,12 @@ import { Box, Typography } from '@mui/material'
 import { BasicInfoProps } from '../../types/types'
 
 const BasicInfo = ({ os, storage, dimension, release }: BasicInfoProps) => {
-  console.log(typeof os)
+  const onlyReleaseDate: string = release.replace(/Released /i, '')
   return (
     <Box>
+      <Typography variant="h4" textAlign={'center'} className="pb-2">
+        Basic Information
+      </Typography>
       <Typography variant="subtitle1">
         <span className="font-semibold">OS:</span> {os}
       </Typography>
@@ -14,7 +17,9 @@ const BasicInfo = ({ os, storage, dimension, release }: BasicInfoProps) => {
       <Typography variant="subtitle1">
         <span className="font-semibold">Dimensions:</span> {dimension}
       </Typography>
-      <Typography variant="subtitle1">{release}</Typography>
+      <Typography variant="subtitle1">
+        <span className="font-semibold">Release Date:</span> {onlyReleaseDate}
+      </Typography>
     </Box>
   )
 }
